Cherry-pick #17753 to 7.x: Fix prometheus histogram rate overflows (#17783)
Fix some overflows on Prometheus histogram rate calculations.

They could be caused by:
* New buckets added to existing histograms on runtime, this happens at
  least with CockroachDB (see #17736).
* Buckets with bigger upper limits have lower counters. This is wrong and
  has been only reproduced this on tests, but handling it just in case to avoid
  losing other data if this happens with some service.

Rate calculation methods return now also a boolean to be able to differenciate
if a zero value is caused because it was the first call, or because it the rate is
actually zero.

(cherry picked from commit 0afffa8e1e267c745ea949fd2dfeb70bfc5f436f)